多次改版迭代开发

+++ src/main/java/com/hnu/ict/ids/entity/TravelTicketInfo.java
@@ -25,7 +25,7 @@ public class TravelTicketInfo {
     /**
      * 乘车用户id
      */
-    private BigInteger userId;
+    private Integer userId;
 
 
     /**
